+++ .idea/dataSources.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="DataSourceManagerImpl" format="xml" multifile-model="true">
-    <data-source source="LOCAL" name="real sqlite" uuid="7e7b4b83-c521-4c3e-8fd2-e549e737f332">
+    <data-source source="LOCAL" name="SQLite - db.sqlite" uuid="ac7eebc5-e5b7-4b66-b093-93fdd4bfa4ff">
       <driver-ref>sqlite.xerial</driver-ref>
       <synchronize>true</synchronize>
       <jdbc-driver>org.sqlite.JDBC</jdbc-driver>
